Merged PR 9527: Authoring network interface topic
Adding a page per customer feedback on the Drivers repo (issue [#609](https://github.com/MicrosoftDocs/windows-driver-docs/issues/609)). This is a general, high-level conceptual page describing network interfaces and their identifiers in code. It can be linked to by any API that uses such an identifier as a parameter regardless of whether it is a desktop API or a driver DDI.