Merged PR 12954: Updated samples on Using SQL and AQS to Query the Index. Also fixed a number of
Updated samples on Using SQL and AQS to Query the Index. Also fixed a number of linting errors that has slipped in as well.

Let me know if there are other people that I should include on the review as well. All the changes are to the code sample, no changes to the text