Merged PR 13586: Removed references to Vista as the Current Version
Made a few mechanical updates including:
- Removing references to Vista as the current version
- Updated a few references to 7 as the current version
- Fixed bad escaping that made the XML docs unreadable
- Removed inline HTML in some entries, added annotations where it wasn't possible

Removing all the references to Windows 7 as the current version will take a refactoring of the entire doc set, so that is left as a future exercise. As well all the articles have inline HTML and linting issues, so I'm only fixing the ones that we are making other changes to. Older articles will be fixed as time allows.

Note there were some build errors about bad links, but I think those are noise. The links work on my machine and on the live site.