Merged PR 13663: Updated sample links to point to GitHub
Updated any references to Code Gallery or the Windows 7 SDK to point to the correct locations on GitHub.

Also including a warning that the solution files will need upgrading on first run if you are using a newer version of Visual Studio.

Also removed some of the less useful build instructions, as Windows development has really coalesced around Visual Studio since the XP/Vista time frame.

Note that almost all the files in the search doc set have linting errors that have crept in over the years, I'm only fixing them in the files that are being touched. Hopefully that will get most of them as the docs are updated over time.