Fix broken link to BasicHLSL sample
This fixes a link to point to a suitable substitute, MSDN sample code for BasicHLSL from Chuck Walbourn, instead of 404. 
Changed words to "based on" because the MSDN sample code uses tex2D instead of Sample.
The substitute is for the 11 version of the sample but it doesn't matter in this context.