Correct export directory example code
The example code is wrong and does not work, see [this](https://stackoverflow.com/a/20453007) answer from 2013 and [this](https://stackoverflow.com/a/40001778) from 2016. Tried to make the explanation more clear and added an example how to search for an RVA by ordinal instead of by name.
The corrected algorithm yields the same results as `GetProcAddress(name)` and `GetProcAddress(biased_ordinal)` do.