Fix x64 crash due to pointer truncation
The code stores a pointer in the window that points to the DemoApp object instance (using SetWindowLongPtrW). The idea is to retrieve the object pointer later when it is needed for processing window message. Conversion to unsigned long is done by PtrToUlong and this unfortunately truncates the 64-bit pointer to a 32-bit unsigned long before SetWindowLongPtrW sees it. The corrupted pointer is then later used and at that point the program crashes on most but not all machines (ASLR?).