Compatibility note for tracelogging provider
I just lost a week of time trying to get tracelogging to work on downlevel Win7.  After trying _many_ things, it turned out that this compatibility notice is in TraceLoggingProvider.h and that you can set a simple directive to change the compilation behavior to target arbitrary Windows SDKs.  A note in the docs would have saved me some time, so I'm leaving it for the next dev.