Very minor update to add TLS to the SSL nomenclature
SSL is the old and familiar term; TLS is the newer and now more correct term. Both are commonly used (still), so we should include both of them on this page. I only updated the first paragraph because (a) it's the most important paragraph (b) we really should include TLS in order to match searches for TLS and (c) it's not actually my technology and I'm not the owner of the page, so I wanted to make as minimal of a change as possible.