Update signtool.md for /verify option
Update to clarify signtool.exe verify option checks *only the EMBEDDED* signature if no catalog options are specified on the command line. Code review of the tool (/ds/security/cryptoapi/pkisign/tools/signtool/*) shows this is correct. However, this is not clear from any of our docs, nor does the tool output indicate that the signature is embedded if it is an embedded signature.