Wrong path to new "Max Path Length..." page in "Naming Files, Paths, and Namespaces"
Path to relocated content in new "Maximum Path Length Limitation" page is incorrect. It points to another directory when the new file is in the same directory as this one. The result is that the user is taken to a non-existent URL and get a 404 error.

I simply removed the path and added **.md** to the filename (in both places) to be consistent with how the link to the "File Streams" page is being handled.

This error was introduced in the following two commits:

* [855d8d3](https://github.com/MicrosoftDocs/win32/commit/855d8d3527f0179e802fd0395f7f50dd8e361524#diff-81fad2a8f974c7b1b74b6c862688e8d7)
* [e19c3f9](https://github.com/MicrosoftDocs/win32/commit/e19c3f9eabe9103c1e5aca7f97ccd9d41320c2d7)