Fix TIFF codec option docs
Formatted the TIFF codec options as a table.

I've also improved the wording of the `CompressionQuality` option description.  The default value for this option is currently missing from the docs.  I attempted to determine the actual default value through experimentation with the Windows 10 version of the codec, but no value made any difference in file size, so I've given a default of `0` to reflect the default value returned by `IPropertyBag2::Read`.

And I have changed the casing of the `TiffCompressionMethod` name to match what's returned by `IPropertyBag2::GetPropertyInfo` from the Windows 10 version of the codec.