Update CONTRIBUTING.md
Hi,

My edit is not actually to remove these entries, however I could not find a way to create an issue against this repository. Apologies in advance if this is not an appropriate way to provide feedback.

As a non-microsoft person, I cannot access either of these links:

https://review.docs.microsoft.com/windows-authoring-guide/?branch=master
https://review.docs.microsoft.com/help/contribute/contribute-how-to-write-pull-request-quality-criteria?branch=master

I get a message saying:

> We are sorry, the page is forbidden.
> You do not have permission with this credential, please sign out to try with others and go back again.

Could the contents be made publicly available? Or at least the relevant parts added inline? If not, perhaps the links should be removed since contributors are likely to want to follow any rules/guidelines and them being unavailable might deter people from bothering.

Thanks.