Update enumerating-all-processes.md
Hi, I was trying to use this code and most of my applications showed "<unknown>". I was wondering if we could add a note about x64/x64 mode applications. I'm really new to trying to use C++ so I am not sure if this is something that was supposed to be very obvious. Please let me know if there is a better way to do this!