Add link from 'Raytracing HLSL Reference' page to State Objects page
There's this hub page linking to all the places you can find DXR HLSL syntax. 
In the interest of completeness, hub page should also link to 'State objects', since you can also specify DXR stuff through those.
It's formatted as a link rather than a row in the table because the linked page is not technically part of this section.