Update SDDL tables (long overdue)
There have been many additions to SDDL that went undocumented.
Fill in all the changes since Windows Vista.

Also, many existing SDDL entries were not marked with the
supported OS versions (if not supported on Windows Server 2003,
which appears to be our baseline). Added markings.

It seems that our policy for marking entries is to list the
versions of Windows they are *not* supported on, which makes
for some very awkward descriptions. (I mean, check out
`SDDL_CRITICAL`!) For "RM" I didn't even try,
because every time we release a new OS, we would have to update
the entry for "RM" and that is not tenable. So "RM" breaks
the rule and instead lists the OS versions that support it (exactly one:
Windows Vista).

It was hard to find descriptions for some of the new entries.
I tried to reuse them from [Well-known SIDs](https://docs.microsoft.com/windows/win32/secauthz/well-known-sids).
Failing that, I used the comment in sddl.h.