Message fixes: Header and duplicate description validation 3 [BULK] (#1022)
* Header validation.

* Consistent casing (at least, within the table)

The first two words of the topic imply that "Render states" is the correct sentence-casing. So I just tweaked that table header. We still have endemic issues across the docs with incorrect casing (for example, the "Render States (Direct3D 9)" topic title), but we're fixing that kind of thing as we go. :)

Co-authored-by: Steven White <stwhi@microsoft.com>