Update "Direct3D feature levels" page to describe feature level 12_2 (#1095)
* Update "Direct3D feature levels" page to describe feature level 12_2

Feature level 12_2 has been out and public for some time. Although it is still in Windows Insider Preview SDKs, it is not going to change (at least, not in any way that could affect information on this page). To make developers' lives easier, this page should be kept up to date.

* Edit

* Fix 12.2 section headers to have the right footnote superscript

* Update overviews-direct3d-11-devices-downlevel-intro.md

* Update header

Co-authored-by: Steven White <stwhi@microsoft.com>