Updated MCN_GETDAYSTATE topic
NMDAYSTATE->prgDayState array is pre-allocated on the stack by the sender of this notification (calendar.cpp). Developer who handles this notification should not be allocating a new array that provides bitfields with bold dates, they should set bits in the given array.  See this topic for the correct handling of this notification: https://docs.microsoft.com/en-us/cpp/mfc/setting-the-day-state-of-a-month-calendar-control?view=msvc-160