Task 36552572: Edit signtool docs /fd /td info
Changes
Top Note: "A warning (error code 0, initially) will be thrown if /fd is not specified during signing and if /td is not specified during timestamping. In later versions of SignTool, the warning will become an error. " --> An error (error code 1) will be thrown if /fd is not specified during signing and if /td is not specified during timestamping.
All table rows where /td specified the note in the row needs to be updated to "An error is generated if /td is not provided while timestamping." (there is no default anymore <- does not need to go in doc"
All table rows where /fd is specified the note in the row needs to be updated to "An error  is generated if the/fd switch is not provided while signing."