Update wm-measureitem.md
`WParam` description in https://docs.microsoft.com/en-us/windows/win32/controls/wm-drawitem#parameters is accurate, it says ` If the message was sent by a menu, this parameter is zero.` WParam in `WM_MEASUREITEM` is used in the same way, however description says the reverse - `If the value is zero, the message was sent by a menu.` This is not correct. `WParam` value will also be set to zero if the combo or list box'control's resource ID is zero.