Update dynamic-link-library-best-practices.md
Removing the "reason" after the rule that says don't call registry functions regarding advapi32. We are getting a lot of developers that read the reason and try to mitigate the reason through various means. This reason is only one of many reasons you cannot use registry functions in dllmain, e.g., if you use a pre-defined handle the registry code will use a critical section which can lead to deadlocks.