Update using-sspi-with-a-windows-sockets-client.md
Running this sample code throws SEC_E_INVALID_HANDLE (0x80090301) error. 

The reason seems to be the improper usage of _SecHandle variables. The global variable hcText is initialized in DoAuthenticate() method, whereas the local variable hCtxt is used in the main() method (which is not initialized). 

Fix : Replaced the local variable hCtxt with the global variable hcText. Also some changes with respect to SspiExample.h. (typecast to GenClientContext parameter to SEC_WCHAR* and MyHandleError parameter to const char*)

TODO: Also please remove the unused _SecHandle parameter from the ConnectAuthSocket() in this file as well as from SspiExample.h.