[BrokenLinksH2] Fix path in link (#1366)
* [BrokenLinksH2] Fix path in link

removing broken link

* Update reading-drm-protected-asf-files-in-directshow.md

I searched for "Microsoft privacy policy" and found the "Microsoft Privacy Statement". I also found other examples in the docs where we talk about the Microsoft Individualization Service and link folks to the Microsoft Privacy Statement. So I did the same thing here. I think it's good to provide the customer with an updated link when we can.

Co-authored-by: Steven White <31261191+stevewhims@users.noreply.github.com>