[BrokenLinksH2] Fix path in link (#1356)
* fixing broken links

* fixing broken link

* Update boundingfrustum-ctor.md

* Update what-s-new-in-directwrite-for-windows-8-consumer-preview.md

* Update what-s-new-in-directwrite-for-windows-8-consumer-preview.md

* Update what-s-new-in-directwrite-for-windows-8-consumer-preview.md

* Corrections

The updated urls were correct, but the link text was still incorrect. And the description was incorrect. Fixed.

* Link text

* Corrections

This is another case where we can't just change the url to point to a different API from the one that was intended. DCompositionGetStatistics and DCompositionGetTargetStatistics are two different APIs (as you can see if you look at the stubs here https://cpubwin.visualstudio.com/win32/_git/sdk-api?path=/sdk-api-src/content/dcomp/nf-dcomp-dcompositiongetstatistics.md&version=GBstubs/main&_a=contents). So we if we're meant to be linking to one, we can't use the url of the other.

I've reverted the changes to this file. The fix is to get the topic published. The topic does exist (it's here https://cpubwin.visualstudio.com/win32/_git/sdk-api?path=/sdk-api-src/content/dcomp/nf-dcomp-dcompositiongettargetstatistics.md), but it's not being published.

* Fixing my previous commit

Co-authored-by: Steven White <31261191+stevewhims@users.noreply.github.com>