Update reporting-an-event.md
Line 115: new line needed for mc compiler

Line 288-293: rhs is a const w_char* but lhs was LPWSTR which is w_char*. It needs to be LPCWSTR. Otherwise code could not be compiled. Hence pInsertString have to be LPCWSTR array. 

Line 312: ReportEvent last argument needs LPVOID but LPWCSTR is not suitable. Const_cast needed.