Fixed various
Some parts of the article referred to lpReserved, while others used lpvReserved. Need to be consistent for clarity. Also updated sample code to make sure the process detach checked the lpvReserved and if it was not NULL to avoid all cleanup  code (as detailed in the remarks of the artlcle)