Update description of IP_MULTICAST_LOOP
I have been working on a [multicast library for Python](https://github.com/multiplemonomials/multicast_expert), and have done some testing with different multicast options on different platforms.  

I found the current description of IP_MULTICAST_LOOP to be incredibly misleading because it doesn't specify whether it needs to be set on the transmitting socket, the receiving one, or both.  I tested it, and it only works on receiving sockets; it does nothing on transmitting ones.  Additionally, if your sockets are open on the loopback interface, setting the option to false makes the loopback not work, a fact which seems worth mentioning.  

I tried to write a new description based on the behavior I've seen when testing this option.  Maybe someone with access to the Windows source could check that I'm right, but I'm fairly sure of this behavior from my testing.  Hope this makes sense!