Update wm-displaychange.md
This documentation needs to be updated to include multiple-display setups. For a single-display system, the usage is clear. For multi-display systems, there's no obvious way to relate this message's **width**, **height** and **bpp** with a _specific_ display on the system. Since multiple displays can have identical modes set, it's not clear how to use this information effectively.