Installation source is listed incorrectly (#1357)
The help document for this page names the installation source for orca.exe by an old name, orca.msi.  Current versions of the SDK source have changed the installation package name.  This makes installation more difficult as it is harder to locate.

Update the document to report a suggested name or the localized naming convention used to help developers locate the installer.