Update OLE/COM Object Viewer doc for 2022
Fixes per conversation with Marian Luparu and Scott Hanselman:
The linked article in the C++ docs repo was deleted in 2017, and the archive link here is broken and shows no content. Since I don't have an easy way to fix the archive, I've moved the original content here, suitably updated for 2022.