Other not interoperate func should be GSS_Unwrap.
I think this paragraph:
Be aware that MakeSignature and VerifySignature will not interoperate with GSS_Wrap when conf_flag is set to zero, or with GSS_Wrap. 
should be changed to:
Be aware that MakeSignature and VerifySignature will not interoperate with GSS_Wrap when conf_flag is set to zero, or with GSS_Unwrap.

The last word for another function which will not interoperate with MakeSignature and VerifySignature should be GSS_Unwrap.