Fix return statement placement in WindowProc
This commit corrects the placement of the "return 0;" statement inside the WM_PAINT case of the WindowProc function. Previously, the return statement was positioned outside the WM_PAINT case's scope, causing it always to return 0 even if the message was not processed. By moving the return statement inside the WM_PAINT case's scope, the function now correctly returns 0 only if the WM_PAINT message is processed. Unhandled messages will be forwarded to the default window procedure as expected.