Fix the example of appending one file to another
`LockFIle` is documented as requiring either `FILE_GENERIC_READ` or `FILE_GENERIC_WRITE` but the example only had `FILE_APPEND_DATA`. This is fixed by using `FILE_APPEND_DATA | FILE_GENERIC_READ`.

Also the example was inconsistent in the case used for the file name. Modern Windows filesystems, in some configurations, can be case sensitive so it's wise to be consistent.