Add two new useful modes to query for HDR support
The previous documented methods only allowed to query HDR support only if it was currently engaged. This is very limiting for games that want to show an HDR/SDR toggle and only grey it out if HDR is not supported at all.
A method to enable HDR from the an app has also been documented, given it's often requested by developers.