Clarify that the 16KB boundary for LFH is only approximate (#1968)
The heap team reserves the right to change the cutoffs for LFH
in the future. Even in the current implementation, the cutoff
is not always exactly at 16KB. Depending on the processor architecture,
heap debugging options, and other external forces, the maximum
size of an LFH block will vary at some value less than or equal
to 16KB.

Added weasel words to the documentation to leave us wiggle room.