Remarks on UB *MemoryBarrierWithGroupSync function
Previously in FXC, AllMemoryBarrierWithGroupSync, DeviceMemoryBarrierWithGroupSync and GroupMemoryBarrierWithGroupSync
would catch divergent thread usage and error out. Now in DXC this is no longer the case as concepts like dynamic uniform exist.
It's important to inform developers of this undefined behavior and it can be tricky to discover if even the docs don't warn you.