Update dx-graphics-hlsl-packing-rules.md
I don't understand the example as given and believe it is written in error.
The documentation earlier in this page states 'every element in an array is stored in a four-component vector.' So why is the problematic example stated as float4[16] when that already is a completely packed array of 64 floats? Also, the modification isn't clear - why is it declaring a 'static' array and trying to cast the original array? float2[32] contains enough space for 128 floats due to each float2 getting padded out to the size of a float4. I think it is clearer to simply offer a rewritten definition of the array.