Update semantic table for HLSL
The current HLSL compiler (DXC) does not handle backward compatibility
the way the documents outlines it.

COLOR is not considered to be system semantic, and POSITION is not
expected to be either.

As we want to fix those inconsistencies at the language/compiler level,
we might want to update this documentation which is the main one showing
up when searching for HLSL semantics on a search engine.

Signed-off-by: Nathan Gauër <brioche@google.com>