Clarify meaning of api-* prefix
Previous text said that "api-*" API sets are "guaranteed to exist on all Windows versions", which is clearly incorrect. An API set introduced in Windows 11 does not exist on Windows 10.

Changed to read "exist on all Windows editions that satisfy the API's version requirements." Got rid of the word "guaranteed" because it adds no value.