winsock: clarify the usage of BSD socket errors
- Clarify that the two style examples would also need the redefinition
  of BSD socket error constants to WSA error constants in order to work
  correctly.

The documentation says BSD socket error constants can be redefined to be
the same as WSA error constants for compatibility, but it was not clear
to me that it applies to the two examples "BSD Style" and
"Preferred Style".

Also, the documentation says the redefinition is highly discouraged so
I would consider renaming so-called "Preferred" style to something else.

Closes #xxxx