Merge latest changes in win32-pr from main to live (#2915)
* Fixed bug in example in 3d-perspective-transform.md

Changes in the example to use Vector2F instead of Vector3F.

* Update cng-mlkem-examples.md (#2910)

The description in the first two paragraphs was incorrect, referring to signing and verifying digital signatures. This is what ML-DSA is used for but not ML-KEM. Corrected the first intro paragraph and combined the second two.

* Fix broken link in VBS Enclaves dev guide (#2914)

Fixing broken link to EnclaveCopyOutOfEnclave fn in API ref docs.

* Update kernel-checks-for-non-whql-signed-drivers.md (#2616)

updating Sysdev to Hardware Developer Program (HDC)

---------

Co-authored-by: learn-build-service-prod[bot] <113403604+learn-build-service-prod[bot]@users.noreply.github.com>
Co-authored-by: nikolaycode <46713216+nikolaycode@users.noreply.github.com>
Co-authored-by: Jess Krynitsky <jekrynit@microsoft.com>
Co-authored-by: Learn Build Service GitHub App <Learn Build Service LearnBuild@microsoft.com>
Co-authored-by: Steven White <31261191+stevewhims@users.noreply.github.com>
Co-authored-by: Cymoki <Cymoki@Microsoft.com>