The FAT describes the cluster chain of the root directory
The FAT describes the cluster chain of the root directory, I assume.

Given that the root directory does not have a file directory entry, there is no record in the ExFAT file system of whether the cluster chain of the root directory is (1) a contiguous cluster chain, or (2) a cluster chain described by the FAT.

I prefer not to assume, so I'm suggesting to document this fact in the specification. Please let me know if I missed something.