RPC tutorial: Switch to unicode Win32 API
Visual Studio 2022 defaults to build against the unicode Win32 API (with "W" suffices). The current sample code assumes the ASCII Win32 (with "A" suffices), which means that it fails to compile in a Visual C++ project with default project settings.

Propose to fix the problem by updating the code to instead use the unicode Win32 API with "W" suffices. It could also be possible to instead fix the problem by adding "A" suffices to the same functions.