RPC tutorial: Link to PRC library directly from sources
Done to make the sample projects to build out-of-the-box without having to tweak linker settings for the project.

It might be that you'll want to clean up or potentially remove https://learn.microsoft.com/en-us/windows/win32/rpc/compiling-and-linking if merging this PR, since the PR more or less makes this page redundant.