Update dot11-phy-type.md
Fixed three problems.
Firstly, there's a format issue for the <dt> values starting about the _vht entry. For Markdown to work, there must be a blank line after the <dt> and before the set of <span> values.

Secondly, the definitions of several of the more recent Wi-Fi types was TBD. I've added definitions that match the rest of the list.

Lastly, some of the definitions presented the information inconsistently. These were changed so that each definition uses the same "template" and mentions which standard (e.g., "802.11ad") and version (e.g., "Wi-Fi 6")