Clarify threading model for Win32 windows
Win32 windows are not "inherently" apartment threaded. You can use Win32 without COM at all, However, *if* a thread uses both COM and Win32 windows, it is expected that that the thread initialize COM as a single-threaded apartment.

(Also, threads are not "apartment-threaded". The term "apartment-threaded" applies to objects, not threads.)